+++ age3-stream-agents/src/main/java/pl/edu/agh/age/compute/stream/configuration/Configuration.java
@@ -28,8 +28,6 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.inject.Inject;
-
 /**
  * Configuration for Stream Agents computation.
  */