+++ age3-stream-agents/src/main/java/pl/edu/agh/age/compute/stream/configuration/WorkplaceConfiguration.java
@@ -34,8 +34,6 @@ import com.google.common.base.MoreObjects;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.inject.Inject;
-
 /**
  * Configuration of a single workplace.
  *